Return correct pointer type from mallocOrExit()etc
Cast the return values from mallocOrExit() etc to the specified pointer
type, so the compiler can check agreement between the <type> argument
and the pointer into which the result is stored.  I didn't do this
originally as it can't be done in general (hence the new need for
mallocOrExitWithoutCast() etc), but the error prevention opportunity
outweighs the inconvenience of needing an extra macro for these unusual
cases (which, in this code base, we'll likely never see).

Also fix the immediate fallout from this newly enabled warning,
one instance of which in fact leads to a crash on 32-bit systems.